+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 [unreleased]
 
 * Add `selectDate` test helper
-* Improve usability of date picker with keyboard
+* Improve usability of date-picker with keyboard
+* Refactor time-picker to be more keyboard friendly
+* Introduce new test helpers: `import { selectTime, getSelectedTime } from 'ember-date-components/test-support/helpers/time-picker'` and `import { selectDate, getSelectedDate, selectDateTime } from 'ember-date-components/test-support/helpers/date-picker'`
 
 v1.2.3